requirements.txt
numpy
pandas
scikit-learn
tqdm
rich
xgboost
lightgbm; platform_system != "Darwin"
pyarrow
xmltodict
pymupdf
Pillow
opencv-python
albumentations
polars
catboost
pyyaml
exa-py
firecrawl-py
sentence-transformers
python-dotenv
kaggle
kagglehub
nltk
transformers[torch]
torchvision
tensorflow
accelerate
peft
bitsandbytes
tf-keras
sentencepiece
wandb
weave
timm
datasets
tensorboard
matplotlib
seaborn
torchgeo
scikit-image
google-genai
